+++ src/codec_aom.c
@@ -43,6 +43,13 @@
 #endif
 #endif
 
+// Speeds 7-9 were added to all intra mode in https://aomedia-review.googlesource.com/c/aom/+/140624.
+#if defined(AOM_EXT_PART_ABI_VERSION)
+#if AOM_ENCODER_ABI_VERSION >= (10 + AOM_CODEC_ABI_VERSION + AOM_EXT_PART_ABI_VERSION)
+#define ALL_INTRA_HAS_SPEEDS_7_TO_9 1
+#endif
+#endif
+
 struct avifCodecInternal
 {
 #if defined(AVIF_CODEC_AOM_DECODE)
@@ -529,17 +536,22 @@ static avifResult aomCodecEncodeImage(avifCodec * codec,
         int aomCpuUsed = -1;
         if (encoder->speed != AVIF_SPEED_DEFAULT) {
             aomCpuUsed = AVIF_CLAMP(encoder->speed, 0, 9);
-#if defined(AOM_USAGE_ALL_INTRA) && \
-    AOM_ENCODER_ABI_VERSION >= (9 + AOM_CODEC_ABI_VERSION + AOM_EXT_PART_ABI_VERSION)
-            // Speeds 7-9 in all intra mode work.
-            if (!(addImageFlags & AVIF_ADD_IMAGE_FLAG_SINGLE) && (aomCpuUsed >= 7)) {
-                aomUsage = AOM_USAGE_REALTIME;
-            }
-#else
             if (aomCpuUsed >= 7) {
+#if defined(AOM_USAGE_ALL_INTRA) && defined(ALL_INTRA_HAS_SPEEDS_7_TO_9)
+                // Speeds 7-9 in all intra mode work.
+                // https://aomedia-review.googlesource.com/c/aom/+/140624
+                // If you were using libaom with lower encoder ABI version
+                // and using the following avif speed setting:
+                //  - speed 0-6: no change is needed
+                //  - speed 7:   change to speed 6 for the same results
+                //  - speed 8-9: re-test and re-adjust speed according to your app needs
+                if (!(addImageFlags & AVIF_ADD_IMAGE_FLAG_SINGLE)) {
+                    aomUsage = AOM_USAGE_REALTIME;
+                }
+#else
                 aomUsage = AOM_USAGE_REALTIME;
-            }
 #endif
+            }
         }
 
         // aom_codec.h says: aom_codec_version() == (major<<16 | minor<<8 | patch)